Update FilterMain.cpp
Here I brought the spot in the color array that the output was pointing to outside of the loop to limit memory fetching. By making it an integer outside of the loop, we don't have to go into memory as much. I also used that in the divisor portion. All of the color references were replaced by output_begin. At the end that color spot in the output was loaded with  output_begin.