Update FilterMain.cpp
got rid of the inner two for loops, added an if statement around the division to hopefully eliminate the division if not necessary. Moved all of the get function calls outside of the loops to limit the amount of calls. each one has a different variable name and is utilized in the loops